feat(lib): create ledger lib and remove methods from transactions lib
